Move language-specific things to a separate file and refactor code
This commit moves most of the constants related to the language to
the new file language.py, with the objective of turning the compiler
into a grammar-independent code. It also changes the token tags from
integers to strings, adds getters and setters for some classes and add
line and column to SyntaxError message.